AND-7105: Add collaboration details to file info screen
- Create custom view to show collaborators in the Invite collaborators fragment.
- Fixed spinner title
