+++ box-share-sdk/src/main/java/com/box/androidsdk/share/fragments/InviteCollaboratorsFragment.java
@@ -12,14 +12,11 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
-import android.widget.LinearLayout;
 import android.widget.MultiAutoCompleteTextView;
-import android.widget.TextView;
 
 import com.box.androidsdk.content.BoxException;
 import com.box.androidsdk.content.BoxFutureTask;
 import com.box.androidsdk.content.models.BoxCollaboration;
-import com.box.androidsdk.content.models.BoxCollaborator;
 import com.box.androidsdk.content.models.BoxFolder;
 import com.box.androidsdk.content.models.BoxIteratorCollaborations;
 import com.box.androidsdk.content.models.BoxUser;
@@ -28,14 +25,13 @@ import com.box.androidsdk.content.requests.BoxResponse;
 import com.box.androidsdk.content.requests.BoxResponseBatch;
 import com.box.androidsdk.content.utils.BoxLogUtils;
 import com.box.androidsdk.content.utils.SdkUtils;
-import com.box.androidsdk.content.views.BoxAvatarView;
 import com.box.androidsdk.share.CollaborationUtils;
 import com.box.androidsdk.share.R;
 import com.box.androidsdk.share.adapters.InviteeAdapter;
 import com.box.androidsdk.share.internal.models.BoxInvitee;
 import com.box.androidsdk.share.internal.models.BoxIteratorInvitees;
 import com.box.androidsdk.share.ui.ChipCollaborationView;
-import com.eclipsesource.json.JsonObject;
+import com.box.androidsdk.share.views.CollaboratorsInitialsView;
 import com.tokenautocomplete.TokenCompleteTextView;
 
 import java.net.HttpURLConnection;
@@ -53,7 +49,6 @@ public class InviteCollaboratorsFragment extends BoxFragment implements View.OnC
 
     private static final Integer MY_PERMISSIONS_REQUEST_READ_CONTACTS = 32;
     protected static final String TAG = InviteCollaboratorsFragment.class.getName();
-    public static final String EXTRA_ACCESS_TOKEN = "InviteCollaboratorsFragment.ExtraAccessToken";
     public static final String EXTRA_USE_CONTACTS_PROVIDER = "InviteCollaboratorsFragment.ExtraUseContactsProvider";
     private Button mRoleButton;
     private ChipCollaborationView mAutoComplete;
@@ -62,19 +57,7 @@ public class InviteCollaboratorsFragment extends BoxFragment implements View.OnC
     private ArrayList<BoxCollaboration.Role> mRoles;
     private InviteCollaboratorsListener mInviteCollaboratorsListener;
     private String mFilterTerm;
-    private TextView mInitialsListHeader;
-    private LinearLayout mInitialsListView;
-    private LinearLayout mInitialsListViewSection;
-    protected BoxIteratorCollaborations mCollaborations;
-    private BoxCollaborator mUnknownCollaborator;
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        JsonObject jsonObject = new JsonObject();
-        jsonObject.add(BoxCollaborator.FIELD_NAME, "");
-        mUnknownCollaborator = new BoxUser(jsonObject);
-        super.onCreate(savedInstanceState);
-    }
+    private CollaboratorsInitialsView mCollabInitialsView;
 
     @Nullable
     @Override
@@ -90,17 +73,12 @@ public class InviteCollaboratorsFragment extends BoxFragment implements View.OnC
         mAdapter.setInviteeAdapterListener(this);
         mAutoComplete.setAdapter(mAdapter);
         mAutoComplete.setTokenListener(this);
-        mInitialsListView = (LinearLayout) view.findViewById(R.id.invite_collaborator_initials_list);
-        mInitialsListHeader = (TextView) view.findViewById(R.id.invite_collaborator_initials_list_header);
-        mInitialsListViewSection = (LinearLayout) view.findViewById(R.id.collaborator_initials_list_section);
-        mInitialsListViewSection.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (mInviteCollaboratorsListener != null) {
-                    mInviteCollaboratorsListener.onShowCollaborators(mCollaborations);
-                }
-            }
-        });
+
+        mCollabInitialsView = (CollaboratorsInitialsView) view.findViewById(R.id.collaboratorsInitials);
+        mCollabInitialsView.setShareItem(mShareItem);
+        mCollabInitialsView.setController(mController);
+        mCollabInitialsView.setInviteCollaboratorsListener(mInviteCollaboratorsListener);
+        mCollabInitialsView.fetchCollaborations();
 
         // Get serialized roles or fetch them if they are not available
         if (getFolder() != null && getFolder().getAllowedInviteeRoles() != null) {
@@ -112,10 +90,10 @@ public class InviteCollaboratorsFragment extends BoxFragment implements View.OnC
         }
 
         fetchInvitees();
-        fetchCollaborations();
         if (getArguments().getBoolean(EXTRA_USE_CONTACTS_PROVIDER)){
             requestPermissionsIfNecessary();
         }
+
         return view;
     }
 
@@ -243,102 +221,6 @@ public class InviteCollaboratorsFragment extends BoxFragment implements View.OnC
      */
     private void fetchInvitees() {
         mController.getInvitees(getFolder(), mFilterTerm).addOnCompletedListener(mGetInviteesListener);
-    }
-
-    /**
-     * Executes the request to retrieve collaborations for the folder
-     */
-    public void fetchCollaborations() {
-        if (getFolder() == null || SdkUtils.isBlank(getFolder().getId())) {
-            mController.showToast(getActivity(), getString(R.string.box_sharesdk_cannot_view_collaborations));
-            return;
-        }
-
-        showSpinner();
-        mController.fetchCollaborations(getFolder()).addOnCompletedListener(mCollaborationsListener);
-    }
-
-    private BoxFutureTask.OnCompletedListener<BoxIteratorCollaborations> mCollaborationsListener =
-            new BoxFutureTask.OnCompletedListener<BoxIteratorCollaborations>() {
-                @Override
-                public void onCompleted(final BoxResponse<BoxIteratorCollaborations> response) {
-                    dismissSpinner();
-                    final Activity activity = getActivity();
-                    if (activity == null) {
-                        return;
-                    }
-                    activity.runOnUiThread(new Runnable() {
-                        @Override
-                        public void run() {
-                            if (response.isSuccess() && getFolder() != null) {
-                                updateUi(response.getResult());
-                            } else {
-                                BoxLogUtils.e(CollaborationsFragment.class.getName(), "Fetch Collaborators request failed",
-                                        response.getException());
-                                mController.showToast(getActivity(), getString(R.string.box_sharesdk_network_error));
-                            }
-                        }
-                    });
-                }
-            };
-
-    public void updateUi(BoxIteratorCollaborations boxIteratorCollaborations) {
-        mCollaborations = boxIteratorCollaborations;
-        if (mCollaborations != null && mCollaborations.size() != 0) {
-
-            mInitialsListHeader.setVisibility(View.VISIBLE);
-            final int totalCollaborators = mCollaborations.fullSize().intValue();
-            final int remainingWidth = mInitialsListView.getWidth();
-            final ArrayList<BoxCollaboration> collaborations = mCollaborations.getEntries();
-
-            clearInitialsView();
-            mInitialsListView.post(new Runnable() {
-                @Override
-                public void run() {
-                    //Add the first item to calculate the width
-                    final View initialsView = addInitialsToList(collaborations.get(0).getAccessibleBy());
-                    initialsView.post(new Runnable() {
-                        @Override
-                        public void run() {
-                            int viewWidth = initialsView.getWidth();
-                            int viewsCount = remainingWidth/viewWidth;
-
-                            for (int i = 1; i < viewsCount && i < collaborations.size(); i++) {
-                                View viewAdded = addInitialsToList(collaborations.get(i).getAccessibleBy());
-                                if (i == viewsCount - 1) {
-                                    // This is the last one, display count if needed
-                                    int remaining = totalCollaborators - viewsCount;
-                                    if (remaining > 0) {
-                                        BoxAvatarView initials = (BoxAvatarView) viewAdded.findViewById(R.id.collaborator_initials);
-                                        JsonObject jsonObject = new JsonObject();
-                                        jsonObject.set(BoxCollaborator.FIELD_NAME, Integer.toString(remaining + 1));
-                                        BoxUser numberUser = new BoxUser(jsonObject);
-                                        initials.loadUser(numberUser, mController.getAvatarController());
-                                    }
-                                }
-                            }
-                        }
-                    });
-                }
-            });
-        }
-    }
-
-    private void clearInitialsView() {
-        mInitialsListView.removeAllViewsInLayout();
-    }
-
-    private View addInitialsToList(BoxCollaborator collaborator) {
-        View layoutContainer =  LayoutInflater.from(getActivity()).inflate(R.layout.view_initials, null);
-        BoxAvatarView initialsView = (BoxAvatarView) layoutContainer.findViewById(R.id.collaborator_initials);
-
-        if (collaborator == null) {
-            initialsView.loadUser(mUnknownCollaborator, mController.getAvatarController());
-        } else {
-            initialsView.loadUser(collaborator, mController.getAvatarController());
-        }
-        mInitialsListView.addView(layoutContainer);
-        return layoutContainer;
     }
 
     /**
