+++ box-share-sdk/src/main/res/layout/fragment_invite_collaborators.xml
@@ -36,28 +36,9 @@
             android:layout_height="@dimen/box_sharesdk_collaboration_cell_height"
             android:layout_marginTop="20dp"
             />
-        <LinearLayout
-            android:id="@+id/collaborator_initials_list_section"
+        <com.box.androidsdk.share.views.CollaboratorsInitialsView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:orientation="vertical"
-            android:layout_marginTop="32dp"
-            android:layout_marginBottom="48dp">
-            <TextView
-                android:id="@+id/invite_collaborator_initials_list_header"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:textAppearance="@style/TextAppearance.AppCompat.Subhead"
-                android:textColor="@color/box_sharesdk_collaborators_color"
-                android:text="@string/box_sharedsdk_collaborators_title"
-                android:visibility="gone"
-                android:layout_marginBottom="10dp"/>
-            <LinearLayout
-                android:id="@+id/invite_collaborator_initials_list"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="horizontal">
-            </LinearLayout>
-        </LinearLayout>
+            android:id="@+id/collaboratorsInitials"/>
     </LinearLayout>
 </ScrollView>
