+++ box-share-sdk/src/main/res/values/dimens.xml
@@ -11,4 +11,5 @@
     <dimen name="box_sharesdk_horizontal_list_item_margin">4dp</dimen>
     <dimen name="box_sharesdk_zero_margin">0dp</dimen>
     <dimen name="box_sharesdk_dialog_margin">24dp</dimen>
+    <dimen name="box_sharesdk_progressbar_size">50dp</dimen>
 </resources>
